+++ core/src/main/scala-3/representation/IR.scala
@@ -5,9 +5,7 @@ import scair.AttrParser
 import scair.Parser
 import scair.Parser.*
 import scair.Printer
-
-import scala.collection.mutable.LinkedHashMap
-import scala.collection.mutable.ListBuffer
+import scair.ir.utils.*
 
 // ██╗ ██████╗░
 // ██║ ██╔══██╗
@@ -15,45 +13,6 @@ import scala.collection.mutable.ListBuffer
 // ██║ ██╔══██╗
 // ██║ ██║░░██║
 // ╚═╝ ╚═╝░░╚═╝
-
-/*≡==--==≡≡≡==--=≡≡*\
-||      UTILS      ||
-\*≡==---==≡==---==≡*/
-
-// TO-DO: export utils to a separate file
-
-val DictType = LinkedHashMap
-type DictType[A, B] = LinkedHashMap[A, B]
-
-val ListType = ListBuffer
-type ListType[A] = ListBuffer[A]
-
-extension (dt: DictType[String, Attribute]) {
-
-  def checkandget(
-      key: String,
-      op_name: String,
-      expected_type: String
-  ): Attribute = {
-    dt.get(key) match {
-      case Some(b) => b
-      case None =>
-        throw new Exception(
-          s"Operation '${op_name}' must include an attribute named '${key}' of type '${}'"
-        )
-    }
-  }
-
-}
-
-extension (lt: ListType[Value[Attribute]]) {
-
-  def updateOperandsAndUses(use: Use, newValue: Value[Attribute]): Unit = {
-    newValue.uses += use
-    lt.update(use.index, newValue)
-  }
-
-}
 
 /*≡==--==≡≡≡≡==--=≡≡*\
 ||    ATTRIBUTES    ||
@@ -179,9 +138,6 @@ class Value[T <: Attribute](
 }
 
 extension (seq: Seq[Value[Attribute]]) def typ: Seq[Attribute] = seq.map(_.typ)
-
-type Operand[T <: Attribute] = Value[T]
-type OpResult[T <: Attribute] = Value[T]
 
 /*≡==--==≡≡≡≡==--=≡≡*\
 ||      BLOCKS      ||
