+++ core/src/main/scala-3/scairdl/Constraints.scala
@@ -2,6 +2,7 @@ package scair.scairdl.constraints
 
 import scair.exceptions.VerifyException
 import scair.ir.*
+import scair.ir.utils.DictType
 
 import scala.reflect.ClassTag
 import scala.util.Failure
