+++ core/src/main/scala-3/transformations/PatternRewriter.scala
@@ -2,6 +2,7 @@ package scair.transformations
 
 import scair.dialects.builtin.ModuleOp
 import scair.ir.*
+import scair.ir.utils.ListType
 
 import scala.collection.mutable.Stack
